core/requirements.txt
appdirs==1.4.4
arrow==1.1.0
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-simple-history==2.12.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
jinja2-time==0.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
Pillow==8.1.0
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
testfixtures==6.17.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
